breeze work

+++ app/Http/Controllers/ReportController.php
@@ -5,12 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Report;
 use App\Models\Status;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ReportController extends Controller
 {
-    public function index(Request $request){
+    public function index(Request $request)
+    {
         $sort = $request->input('sort');
-        if($sort != 'asc' && $sort != 'desc'){
+        if ($sort != 'asc' && $sort != 'desc') {
             $sort = 'desc';
         }
 
@@ -18,39 +20,61 @@ class ReportController extends Controller
         $validate = $request->validate([
             'status' => "exists:statuses,id"
         ]);
-        if($validate){
+        if ($validate) {
             $reports = Report::where('status_id', $status)
-                    ->orderBy('created_at', $sort)
-                    ->paginate(8);
-        }else{
-            $reports = Report::orderBy('created_at', $sort)
-                    ->paginate(8);
+                ->where('user_id', Auth::user()->id)
+                ->orderBy('created_at', $sort)
+                ->paginate(6);
+        } else {
+            $reports = Report::where('user_id', Auth::user()->id)
+                ->orderBy('created_at', $sort)
+                ->paginate(6);
         }
 
         $statuses = Status::all();
         return view('report.index', compact('reports', 'statuses', 'sort', 'status'));
     }
-    public function destroy(Report $report){
-        $report->delete();
-        return redirect()->back();
+    public function destroy(Report $report)
+    {
+        if (Auth::user()->id == $report->user_id) {
+            $report->delete();
+            return redirect()->back();
+        } else {
+            abort(403, 'У вас нет прав на редактирование этой записи.');
+        }
     }
-    public function store(Request $request, Report $report){
-        $data = $request -> validate([
-            'car_number' => 'string|required',
-            'description' => 'string|required',
-        ]);
+    public function store(Request $request, Report $report)
+    {
+            $data = $request->validate([
+                'car_number' => 'string|required',
+                'description' => 'string|required',
+            ]);
 
-        $report->create($data);
-        return redirect()->back();
+            $data['user_id'] = Auth::user()->id;
+            $data['status_id'] = 1;
+
+            $report->create($data);
+            return redirect()->back();
     }
-    public function edit(Report $report){
-        return view('report.edit', compact('report'));
+    public function edit(Report $report)
+    {
+        if (Auth::user()->id == $report->user_id) {
+            return view('report.edit', compact('report'));
+        } else {
+            abort(403, 'У вас нет прав на редактирование этой записи.');
+        }
     }
-    public function update(Request $request, Report $report){
-        $data = $request -> validate([
-            'car_number' => 'string|required',
-            'description' => 'string|required',
-        ]);
+    public function update(Request $request, Report $report)
+    {
+        if (Auth::user()->id == $report->user_id) {
+            $data = $request->validate([
+                'car_number' => 'string|required',
+                'description' => 'string|required',
+            ]);
+        } else {
+            abort(403, 'У вас нет прав на редактирование этой записи.');
+        }
+
 
         $report->update($data);
         return redirect()->back();
